move formatting functions from plot/formatting.ts to labels.ts module
- Move format_ticks and format_tick_label from src/lib/plot/formatting.ts to src/lib/labels.ts
- Delete src/lib/plot/formatting.ts as it's now consolidated into labels module
- Update all imports across components to use labels module instead
- Move formatting tests from tests/vitest/plot/formatting.test.ts to tests/vitest/labels.test.ts
- Update CoordinationBarPlot, ScatterPlot, ScatterPoint, SpacegroupBarPlot, and XrdPlot imports
