+++ src/lib/coordination/CoordinationBarPlot.svelte
@@ -2,8 +2,8 @@
   import type { AnyStructure, BarSeries, BarTooltipProps, Orientation } from '$lib'
   import { plot_colors } from '$lib/colors'
   import { decompress_file, handle_url_drop } from '$lib/io'
+  import { format_value } from '$lib/labels'
   import { BarPlot } from '$lib/plot'
-  import { format_value } from '$lib/plot/formatting'
   import { parse_any_structure } from '$lib/structure/parse'
   import type { ComponentProps } from 'svelte'
   import { SvelteMap, SvelteSet } from 'svelte/reactivity'
