+++ src/lib/plot/index.ts
@@ -13,7 +13,6 @@ export { default as ColorBar } from './ColorBar.svelte'
 export { default as ColorScaleSelect } from './ColorScaleSelect.svelte'
 export * from './data-transform'
 export { default as ElementScatter } from './ElementScatter.svelte'
-export * from './formatting'
 export { default as Histogram } from './Histogram.svelte'
 export { default as HistogramControls } from './HistogramControls.svelte'
 export * from './interactions'
